+++ aerich/migrate.py
@@ -13,15 +13,9 @@ from tortoise import (
     Model,
     Tortoise,
 )
-from tortoise.backends.asyncpg.schema_generator import AsyncpgSchemaGenerator
-from tortoise.backends.mysql.schema_generator import MySQLSchemaGenerator
-from tortoise.backends.sqlite.schema_generator import SqliteSchemaGenerator
 from tortoise.fields import Field
 
 from aerich.ddl import BaseDDL
-from aerich.ddl.mysql import MysqlDDL
-from aerich.ddl.postgres import PostgresDDL
-from aerich.ddl.sqlite import SqliteDDL
 from aerich.exceptions import ConfigurationError
 from aerich.utils import get_app_connection
 
@@ -78,11 +72,17 @@ class Migrate:
         await Tortoise.init(config=migrate_config)
 
         connection = get_app_connection(config, app)
-        if connection.schema_generator is MySQLSchemaGenerator:
+        if connection.schema_generator.DIALECT == "mysql":
+            from aerich.ddl.mysql import MysqlDDL
+
             cls.ddl = MysqlDDL(connection)
-        elif connection.schema_generator is SqliteSchemaGenerator:
+        elif connection.schema_generator.DIALECT == "sqlite":
+            from aerich.ddl.sqlite import SqliteDDL
+
             cls.ddl = SqliteDDL(connection)
-        elif connection.schema_generator is AsyncpgSchemaGenerator:
+        elif connection.schema_generator.DIALECT == "postgres":
+            from aerich.ddl.postgres import PostgresDDL
+
             cls.ddl = PostgresDDL(connection)
         else:
             raise NotImplementedError("Current only support MySQL")