+++ tests/test_migrate.py
@@ -1,5 +1,0 @@
-from tortoise.contrib import test
-
-
-class TestMigrate(test.TruncationTestCase):
-    pass